Expose the helper code as a library so external crate can build the binary themselves
